back/pyproject.toml
[project]
name = "poetry-test3"
version = "0.1.0"
description = ""
authors = [
    {name = "egor",email = "egor.no.login@gmail.com"}
]

requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (==4.0.1)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
